report-management - move get-report into bark
